# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutlinePreviewRequest": {
        "description": "POST /api/v1/projects/outline-preview \u2014 generate a reviewable outline first.",
        "properties": {
          "bpm": {
            "description": "Beats per minute",
            "maximum": 220.0,
            "minimum": 60.0,
            "title": "Bpm",
            "type": "integer"
          },
          "duration_minutes": {
            "default": 5,
            "description": "Target duration in minutes",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "llm_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "default",
            "description": "Server-defined safe LLM profile id",
            "title": "Llm Profile Id"
          },
          "topic": {
            "description": "Podcast topic",
            "maxLength": 200,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          },
          "tts_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Server-defined safe TTS profile id",
            "title": "Tts Profile Id"
          }
        },
        "required": [
          "topic",
          "bpm"
        ],
        "title": "OutlinePreviewRequest",
        "type": "object"
      },
...
      "ProjectCreateRequest": {
        "description": "POST /api/v1/projects \u2014 user intent only; provider secrets are server-side.",
        "properties": {
          "approved_outline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectOutlineResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "User-reviewed outline to use for script generation"
          },
          "bpm": {
            "description": "Beats per minute",
            "maximum": 220.0,
            "minimum": 60.0,
            "title": "Bpm",
            "type": "integer"
          },
          "duration_minutes": {
            "default": 5,
            "description": "Target duration in minutes",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "interviewer_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeakerProfile"
              },
              {
                "type": "null"
              }
            ],
            "description": "Interviewer voice, tone, humor, and style"
          },
          "llm_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "default",
            "description": "Server-defined safe LLM profile id",
            "title": "Llm Profile Id"
          },
          "outline_preview_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque approved outline preview binding",
            "title": "Outline Preview Id"
          },
          "sme_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeakerProfile"
              },
              {
                "type": "null"
              }
            ],
            "description": "Subject matter expert guest voice, tone, humor, and expertise"
          },
          "topic": {
            "description": "Podcast topic",
            "maxLength": 200,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          },
          "tts_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "default",
            "description": "Server-defined safe TTS profile id",
            "title": "Tts Profile Id"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Voice ID from /api/v1/config voices list",
            "title": "Voice Id"
          }
        },
        "required": [
          "topic",
          "bpm"
        ],
        "title": "ProjectCreateRequest",
        "type": "object"
      },
      "ProjectCreateResponse": {
        "description": "Response for POST /api/v1/projects (202 Accepted).",
        "properties": {
          "manifest_url": {
            "title": "Manifest Url",
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/PodcastProjectResponse"
          }
        },
        "required": [
          "project",
          "manifest_url"
        ],
        "title": "ProjectCreateResponse",
        "type": "object"
      },
...
      "ProjectListResponse": {
        "description": "Response for GET /api/v1/projects.",
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "ProjectListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Heavy-lift service for script generation, TTS, beat generation, and audio mixing.",
    "title": "Straub AH \u2014 Podcast Worker Service",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "description": "Stream or download an artifact for the authenticated owner.",
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "v1-artifacts"
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "description": "List durable project manifests for the authenticated owner.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "v1-projects"
        ]
      },
      "post": {
        "description": "Create a durable PodcastProject and start backend generation.",
        "operationId": "create_project_api_v1_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "v1-projects"
        ]
      }
    },
    "/api/v1/projects/outline-preview": {
      "post": {
        "description": "Generate a reviewable outline with a durable paired safe-profile binding.",
        "operationId": "preview_project_outline_api_v1_projects_outline_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlinePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlinePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Project Outline",
        "tags": [
          "v1-projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete the project and its retained artifacts.",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "v1-projects"
        ]
      },
      "get": {
        "description": "Return the canonical backend-owned PodcastProject manifest.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "v1-projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/outline": {
      "get": {
        "description": "Return the project's script outline without full segment text or artifacts.",
        "operationId": "get_project_outline_api_v1_projects__project_id__outline_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOutlineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Outline",
        "tags": [
          "v1-projects"
        ]
      }
    }
  }
}